[arm64] Program PMU counter `4` to track core interrupts/events

+++ aarch64/corefreq.h
@@ -166,7 +166,7 @@ typedef struct
 		} Voltage;
 
 		struct {
-		unsigned int		SMI;
+		unsigned long long	SMI;
 			struct {
 			unsigned int	LOCAL,
 					UNKNOWN,
